Decrease of execution time
When connecting the graph of a document by calling the "connect" method, the execution time could reach several seconds depending by the chosen document. This timing could be proibitive for some applications, especially when dealing with a large number of documents composed by several pages. To reduce the computational time needed by the "connect" method, I have made some modifications on the code dealing with data into the "df" dataframe of the "ObjectTree" class.

I have developed a unit testing code (not attached here) with several tests to verify if the output achieved with the original code was equal to the one achieved with the modified cose. All the tests passed. The modified code was able to achieve the same output in a fraction of seconds in documents for which the original code required several seonds of computation. I hope that this change will help the project.